src/test/6/requirements.txt
NumPy
pytest
Flask
flask-cors
Pillow
json-rpc
requests
transformers
torch
Pillow
accelerate
git+https://github.com/kashif/diffusers.git@wuerstchen-v3
peft
invisible_watermark
safetensors
opencv-python
opencv-python-headless